chore: company page componentized, check tabs for active company
