+++ client/components/App/ConfirmationModal.vue
@@ -7,7 +7,7 @@
             {{ confirmationModalMessage }}
         </div>
         <NBCardFooter>
-            <div class="flex justify-end gap-2 pb-4">
+            <div class="flex justify-end gap-2">
                 <NBButtonSquare @click=handleConfirmButtonClick size="md" class="px-4 hover:bg-secondary-bg">
                     Confirm
                 </NBButtonSquare>
