+++ client/components/App/Dropdown.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script setup>
-import { ref, watch, defineProps, defineEmits } from 'vue';
+
 
 const props = defineProps({
     modelValue: {
